Reorganize files for core logging functionality
